refactor: give renderer preparation and columns clear homes
