+++ src/index.ts
@@ -3,5 +3,5 @@ export * as Columns from "./columns";
 export { Progress } from "./services/progress";
 export * from "./services/stdio";
 export * from "./types";
-export { useNow } from "./services/renderer/context/now-context";
-export { useSpinnerTick } from "./services/renderer/context/spinner-context";
+export { useNow } from "./renderer/context/now-context";
+export { useSpinnerTick } from "./renderer/context/spinner-context";
