+++ src/renderer/context/spinner-context.tsx
