+++ src/renderer/hooks/use-spinner-clock.ts
