+++ src/renderer/shared/text-width.ts
