+++ src/services/progress.ts
@@ -3,7 +3,7 @@ import { adaptTaskApi } from "../task-api";
 import { ProgressStore } from "./store/store";
 import type { AddTaskOptions, ProgressService, TaskHandle, TaskId } from "../types";
 import { Task } from "../types";
-import { Renderer } from "./renderer/renderer";
+import { Renderer } from "../renderer/renderer";
 import { ProgressStdio } from "./stdio";
 
 interface CurrentParentState {
